streamlined tools
I am going to split this tool into parts.
First of all I want a lambda to fetch data from the FPL website and upload it to s3
I then want to build my own api to serve it. I will do this with Strawberry and graphql
This means the CLi will change completely as it will instead talk to the API
There wont be a dynamo table either. So I have removed this code
Also the api will change, so that code is gone too